# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KIN API",
    "version": "1.0.0",
    "description": "KIN AI Companion Platform — We Build You A Friend\n\nThe KIN API powers Mission Control (web dashboard), Telegram/Discord/WhatsApp bots, local KIN instances, and third-party integrations. It provides:\n\n- **Auth** — Telegram Login Widget + JWT-based sessions\n- **Chat** — Multi-companion AI chat with a two-brain supervisor (Local Ollama → Groq → Anthropic/OpenAI fallback)\n- **Skills Marketplace** — Browse, install, and custom-submit companion skill extensions\n- **Companion Skills** — NFT-portable XP/leveling system for companion skill trees\n- **Heartbeat & Recovery** — Local KIN health monitoring and snapshot-based recovery\n- **Billing** — Stripe subscription checkout, portal, and webhook handling\n- **Support** — AI-powered support chat with human escalation, FAQ, tickets, and feature requests\n\nAll protected routes require a JWT bearer token obtained via `POST /auth/telegram` or `POST /auth/dev-login` (development only).",
    "contact": {
      "name": "KIN Support",
      "url": "https://www.meetyourkin.com/support",
      "email": "support@meetyourkin.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/kr8tiv-ai/kr8tiv-runtime-truth-contracts/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://api.meetyourkin.com",
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtain via `POST /auth/telegram` or `POST /auth/dev-login`. Token expires after 2 hours. Algorithm: HS256."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "statusCode": { "type": "integer" }
                }
              }
            ]
          }
        },
        "example": { "error": "Unauthorized" }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "example": { "error": "Bad Request", "message": "Message is required" }
      },
...
      "CompanionId": {
        "type": "string",
        "enum": ["cipher", "mischief", "vortex", "forge", "aether", "catalyst"],
        "description": "Companion identifier. Each companion has a distinct personality and specialisation.",
        "example": "cipher"
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT token.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded. Default: 100 requests/minute per IP in production.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": { "type": "integer", "example": 429 },
                "error": { "type": "string", "example": "Too Many Requests" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict — resource already exists.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  },
  "security": [],
...
  "paths": {
...
    "/billing/status": {
      "get": {
        "tags": ["Billing"],
        "operationId": "billingStatus",
        "summary": "Subscription status",
        "description": "Returns the current subscription plan, status, billing period, and live usage statistics (messages today, active companions, total API calls).",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Billing status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": "string",
                      "enum": ["free", "pro", "enterprise"],
                      "example": "pro"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["active", "past_due", "canceled", "trialing"],
                      "example": "active"
                    },
                    "currentPeriodStart": { "type": "string", "format": "date-time", "nullable": true },
                    "currentPeriodEnd": { "type": "string", "format": "date-time", "nullable": true },
                    "cancelAtPeriodEnd": { "type": "boolean" },
                    "stripeSubscriptionId": { "type": "string", "nullable": true },
                    "usage": {
                      "type": "object",
                      "properties": {
                        "messagesToday": { "type": "integer", "example": 12 },
                        "activeCompanions": { "type": "integer", "example": 2 },
                        "apiCalls": { "type": "integer", "example": 347 }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/billing/checkout": {
      "post": {
        "tags": ["Billing"],
        "operationId": "billingCheckout",
        "summary": "Stripe checkout",
        "description": "Creates a Stripe Checkout session for subscription purchase. Returns the Stripe-hosted checkout URL. If `STRIPE_SECRET_KEY` is not configured, returns `{url: null, message: 'Payments coming soon'}`.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["priceId"],
                "properties": {
                  "priceId": {
                    "type": "string",
                    "description": "Stripe Price ID for the subscription plan.",
                    "example": "price_1OaBcDEfGhIjKlMnOpQrStUv"
                  },
                  "successUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "Redirect URL after successful payment.",
                    "default": "https://www.meetyourkin.com/billing/success"
                  },
                  "cancelUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "Redirect URL on payment cancellation.",
                    "default": "https://www.meetyourkin.com/billing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout URL.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string", "format": "uri", "nullable": true },
                    "message": { "type": "string", "description": "Present only when Stripe is not configured." }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/billing/mint-checkout": {
      "post": {
        "tags": ["Billing"],
        "operationId": "billingMintCheckout",
        "summary": "Companion NFT mint checkout",
        "description": "Creates a Stripe Checkout session for a one-time companion NFT mint ($9.99). On payment success, the Stripe webhook triggers the Candy Machine mint and records NFT ownership. Users do not need any crypto knowledge — they simply pay and the NFT is minted to their wallet.\n\nRate limited to prevent minting abuse.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["companionId", "walletAddress"],
                "properties": {
                  "companionId": { "$ref": "#/components/schemas/CompanionId" },
                  "walletAddress": {
                    "type": "string",
                    "description": "Solana wallet address to receive the minted NFT.",
                    "example": "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
                  },
                  "successUrl": {
                    "type": "string",
                    "format": "uri",
                    "default": "https://www.meetyourkin.com/dashboard?minted=true"
                  },
                  "cancelUrl": {
                    "type": "string",
                    "format": "uri",
                    "default": "https://www.meetyourkin.com/companions"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mint checkout URL.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string", "format": "uri", "nullable": true },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/billing/portal": {
      "post": {
        "tags": ["Billing"],
        "operationId": "billingPortal",
        "summary": "Stripe portal",
        "description": "Creates a Stripe Customer Portal session. Allows users to manage their subscription, update payment methods, view invoices, and cancel. Requires an existing Stripe customer record — the user must have subscribed at least once.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "returnUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL to return to after leaving the portal.",
                    "default": "https://www.meetyourkin.com/billing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stripe portal URL.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string", "format": "uri", "nullable": true },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/billing/webhook": {
      "post": {
        "tags": ["Billing"],
        "operationId": "billingWebhook",
        "summary": "Stripe webhook",
        "description": "Receives and processes Stripe webhook events. Signature is verified using `STRIPE_WEBHOOK_SECRET` (HMAC-SHA256 with 5-minute tolerance). Handles:\n\n- `checkout.session.completed` — Activates subscriptions and processes companion mint payments\n- `customer.subscription.updated` — Syncs subscription status and period dates\n- `customer.subscription.deleted` — Downgrades user to free tier\n\nThe request body must be the raw Stripe JSON payload (not parsed). Register this URL in your Stripe dashboard.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Raw Stripe event payload. Do not transform.",
                "properties": {
                  "id": { "type": "string", "example": "evt_1OaBcDEfGhIjKlMn" },
                  "type": {
                    "type": "string",
                    "example": "checkout.session.completed",
                    "enum": [
                      "checkout.session.completed",
                      "customer.subscription.updated",
                      "customer.subscription.deleted"
                    ]
                  },
                  "data": { "type": "object" }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "stripe-signature",
            "in": "header",
            "required": false,
            "description": "Stripe-Signature header for webhook verification.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Event acknowledged.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "received": { "type": "boolean", "example": true } }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Stripe signature or malformed JSON.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  }
}